==> 04_csv_To_h5ad_conversion.ipynb <==
import numpy as np
import pandas as pd
import scanpy as sc
import matplotlib as mpl
import matplotlib.pyplot as plt
from colorspacious import cspace_converter

sc.settings.verbosity = 3             # verbosity: errors (0), warnings (1), info (2), hints (3)
sc.logging.print_header()
sc.settings.set_figure_params(dpi=80, facecolor='white')

adata = sc.read_csv('/vol/ExtraVol/MergedData_Preprocessed.csv')
adata = adata.T

anno = pd.read_csv('/vol/ExtraVol/MergedData_Preprocessed_Metadata.csv')

adata.obs = anno

adata

adata.var_names

adata.write_h5ad("/vol/ExtraVol/MergedData_Preprocessed.h5ad") 

### FINISH HERE ###

sc.pl.paga(adata, color=['leiden'])

adata.uns['iroot'] = np.flatnonzero((adata.obs['stage'] == 'embryonic'))[0]

sc.tl.dpt(adata)

sc.pl.umap(adata, color=['leiden', 'dpt_pseudotime'], legend_loc='on data')

sc.tl.rank_genes_groups(adata, 'leiden', method='wilcoxon', use_raw = False)
sc.pl.rank_genes_groups(adata, n_genes=25, sharey=False)

sc.tl.rank_genes_groups(adata, 'leiden', method='wilcoxon', use_raw = False, groups = ['15'], reference = '12',)
sc.pl.rank_genes_groups(adata, n_genes=25, sharey=False)

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Adamdec1+_adult_fibroblasts'])

# Define the clusters of interest
clusters_of_interest = {'0'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Adamdec1+_adult_fibroblasts'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Ogn+_adult_fibroblasts'])

# Define the clusters of interest
clusters_of_interest = {'1'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Ogn+_adult_fibroblasts'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Jam2+_cells'])

# Define the clusters of interest
clusters_of_interest = {'18'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Jam2+_cells'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Sox6+Bmp5+Tcf4+_adult_fibroblasts'])

# Define the clusters of interest
clusters_of_interest = {'7'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Sox6+Bmp5+Tcf4+_adult_fibroblasts'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Adamdec1+_AW_fibroblasts'])

# Define the clusters of interest
clusters_of_interest = {'5'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Adamdec1+_AW_fibroblasts'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Ogn+_AW_fibroblasts'])

# Define the clusters of interest
clusters_of_interest = {'4'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Ogn+_AW_fibroblasts'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Ncam1+_newborn_cells'])

# Define the clusters of interest
clusters_of_interest = {'2'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Ncam1+_newborn_cells'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Jam2+_newborn_cells'])

# Define the clusters of interest
clusters_of_interest = {'8'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Jam2+_newborn_cells'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Igf2+_newborn_cells'])

# Define the clusters of interest
clusters_of_interest = {'9'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Igf2+_newborn_cells'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Colec10+_embryonic_cells'])

# Define the clusters of interest
clusters_of_interest = {'3', '10'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Colec10+_embryonic_cells'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Sox6+Bmp5+Tcf4+_embryonic_cells'])

# Define the clusters of interest
clusters_of_interest = {'17'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Sox6+Bmp5+Tcf4+_embryonic_cells'

adata.obs['cell_type'] = adata.obs['cell_type'].cat.add_categories(['Igf2+_embryonic_cells'])

# Define the clusters of interest
clusters_of_interest = {'6','12','15'}

# Use the isin method to create a boolean mask of the cells that are in the clusters of interest
mask = adata.obs['leiden'].isin(clusters_of_interest)

# Update the 'cell_type' column based on the mask
adata.obs.loc[mask, 'cell_type'] = 'Igf2+_embryonic_cells'

sc.pl.umap(adata, color=['cell_type'])

adata.write_h5ad("/vol/ExtraVol/Development_CellTypeAssigned.h5ad")
